feat: split logos chat crate
