+++ bin/chat-cli/Cargo.toml
@@ -11,7 +11,7 @@ path = "src/main.rs"
 # Workspace dependencies (sorted)
 crossbeam-channel = { workspace = true }
 logos-account = { workspace = true, features = ["dev"] }
-logos-chat = { workspace = true, features = ["embedded-logos-delivery"] }
+logos-chat = { workspace = true }
 
 # External dependencies (sorted)
 anyhow = "1.0"
