+++ crates/generic-chat/examples/message-exchange/README.md
@@ -1,9 +1,9 @@
 # message-exchange
 
-An example Rust application built on top of [`crates/client`](../../).
+An example Rust application built on top of [`crates/generic-chat`](../../).
 
 It demonstrates that creating a working chat client in pure Rust is trivial: depend on
-`crates/client`, pick a `DeliveryService` implementation (here the in-memory
+`crates/generic-chat`, pick a `DeliveryService` implementation (here the in-memory
 `InProcessDelivery` shipped with the crate), and wire up `ChatClient`. No boilerplate, no FFI.
 
 ## Running
