+++ crates/generic-chat/src/builder.rs
