+++ crates/generic-chat/src/client.rs
