+++ crates/generic-chat/src/delegate.rs
