+++ crates/generic-chat/src/errors.rs
