+++ crates/generic-chat/src/event.rs
