+++ crates/logos-chat/src/logos.rs
@@ -1,26 +1,27 @@
 //! The opinionated Logos client.
 //!
-//! `LogosChatClient` commits to the Logos service stack so independently built
-//! clients share the same production services instead of each re-deriving them:
-//! a delegate identity, the HTTP keypackage + account registry, and encrypted
-//! on-disk storage. The stack is generic over the transport — any
-//! [`Transport`] can be injected via
-//! [`open_with_transport`](LogosChatClient::open_with_transport) — and
-//! concrete Logos clients are named aliases that commit to one, each in its
-//! own module (e.g. `EmbeddedLogosClient` in `crate::embedded`, behind the
-//! cargo feature carrying its native dependency). This module itself compiles
-//! unconditionally.
+//! `LogosChatClientInternal` commits to the Logos service stack so
+//! independently built clients share the same production services instead of
+//! each re-deriving them: a delegate identity, the HTTP keypackage + account
+//! registry, and encrypted on-disk storage. The stack is generic over the
+//! transport — any [`Transport`] can be injected via
+//! [`LogosConfig::open_with_transport`] — and concrete Logos clients are
+//! named aliases that commit to one, each in its own module (e.g.
+//! [`LogosChatClient`](crate::LogosChatClient) in `crate::embedded`).
+//!
+//! The client aliases point at `ChatClient`, which lives in
+//! `logos-generic-chat`, so Rust's inherent-impl rule keeps the open
+//! constructors off the aliases; they hang on [`LogosConfig`] (owned here)
+//! instead.
 
 use components::HttpRegistry;
 use crossbeam_channel::Receiver;
 use libchat::{ChatStorage, StorageConfig};
 use logos_account::TestLogosAccount;
 
-use crate::ChatClientBuilder;
-use crate::client::{ChatClient, Transport};
-use crate::delegate::DelegateSigner;
-use crate::errors::ClientError;
-use crate::event::Event;
+use logos_generic_chat::{
+    ChatClient, ChatClientBuilder, ClientError, DelegateSigner, Event, Transport,
+};
 
 /// The endpoint for the account and keypackage registration service.
 pub const REGISTRY_ENDPOINT: &str = "https://devnet.chat-kc.logos.co";
@@ -56,31 +57,20 @@ impl LogosConfig {
     pub fn set_registry_url(&mut self, registry_url: impl Into<String>) {
         self.registry_url = registry_url.into();
     }
-}
 
-/// A [`ChatClient`] wired to the Logos service stack: a [`DelegateSigner`]
-/// identity acting for a fresh dev account, the HTTP keypackage + account
-/// registry ([`HttpRegistry`], which is both the keypackage store and the
-/// account → device directory), and encrypted [`ChatStorage`] — generic over
-/// the transport `T`. Commit to a transport with
-/// [`open_with_transport`](Self::open_with_transport), or use a concrete
-/// alias like `EmbeddedLogosClient`.
-pub type LogosChatClient<T> = ChatClient<T, HttpRegistry, ChatStorage>;
-
-impl<T: Transport> LogosChatClient<T> {
-    /// Open a client on the Logos stack per `config` with the injected
+    /// Open a client on the Logos stack per this config with the injected
     /// transport, persisting to the encrypted database.
-    pub fn open_with_transport(
-        config: LogosConfig,
+    pub fn open_with_transport<T: Transport>(
+        self,
         transport: T,
-    ) -> Result<(Self, Receiver<Event>), ClientError> {
+    ) -> Result<(LogosChatClientInternal<T>, Receiver<Event>), ClientError> {
         // A fresh account endorsing a fresh delegate each open: the account
         // key is dropped after publishing the bundle, so devices cannot be
         // added later. A caller-supplied, custody-holding account replaces
         // this once the platform provides one.
         let account = TestLogosAccount::new();
         let delegate = DelegateSigner::random();
-        let mut registry = HttpRegistry::new(config.registry_url);
+        let mut registry = HttpRegistry::new(self.registry_url);
         account
             .add_delegate_signer(&mut registry, delegate.public_key())
             .map_err(|e| ClientError::BundlePublish(e.to_string()))?;
@@ -89,9 +79,18 @@ impl<T: Transport> LogosChatClient<T> {
             .transport(transport)
             .registration(registry)
             .storage_config(StorageConfig::Encrypted {
-                path: config.db_path,
-                key: config.db_key,
+                path: self.db_path,
+                key: self.db_key,
             })
             .build()
     }
 }
+
+/// A [`ChatClient`] wired to the Logos service stack: a [`DelegateSigner`]
+/// identity acting for a fresh dev account, the HTTP keypackage + account
+/// registry ([`HttpRegistry`], which is both the keypackage store and the
+/// account → device directory), and encrypted [`ChatStorage`] — generic over
+/// the transport `T`. Commit to a transport with
+/// [`LogosConfig::open_with_transport`], or use a concrete alias like
+/// [`LogosChatClient`](crate::LogosChatClient).
+pub type LogosChatClientInternal<T> = ChatClient<T, HttpRegistry, ChatStorage>;
